requirements.txt
aerich==0.7.1
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==22.2.0
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
colorthief==0.2.1
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
iso8601==1.1.0
loguru==0.6.0
multidict==6.0.4
Pillow==9.4.0
pycparser==2.21
PyNaCl==1.5.0
pypika-tortoise==0.1.6
pytz==2022.7.1
tortoise-orm==0.19.2
typing_extensions==4.4.0
win32-setctime==1.1.0
yarl==1.8.2
